requirements.txt
dash==3.0.4
dash-bootstrap-components==2.0.4
dash-daq==0.6.0

pandas>=2.0,<3.0
plotly>=5.20,<6.0
pycountry>=22.3
pyyaml>=6.0
pyarrow>=14.0

urllib3<2